refactor(habitat): move hook model into module
