+++ tools/habitat-harness/src/service/modules/hook/model/dto/hook.schema.ts
