+++ tools/habitat-harness/src/service/modules/hook/model/policy/check-command.policy.ts
