+++ tools/habitat-harness/src/service/modules/hook/model/policy/resource-inspection.policy.ts
