+++ tools/habitat-harness/src/service/modules/hook/model/policy/staged-worktree.policy.ts
